Add Minimap layers to all scenes.
Add more Quest Indicators for Quests 2, 4, 5, 6 and 7.
Edit Minimap icons for Player, Enemies and NPCs.
