+++ Shadow/Assets/Art/Tilesets/MinimapPalette.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:112d8497c95ece0b892d4a5820b67e2edd03fb2a4b8b76502ba673c3912ec534
-size 442
+oid sha256:b4e922d5a8d1d9b37b3335c3b008ddd8804f449af65732b66f2ce5e73da0c593
+size 797
