+++ Shadow/Assets/Prefabs/CharacterClasses/Sorcerer.prefab
@@ -34,6 +34,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 7021806161379852792}
+  - {fileID: 4059791866630638263}
   - {fileID: 537751579799166602}
   m_Father: {fileID: 0}
   m_RootOrder: 0
@@ -237,6 +238,81 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 0.9, y: 0.9}
   m_EdgeRadius: 0
+--- !u!1001 &4011222115521916115
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 7021806160790640475}
+    m_Modifications:
+    - target: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 0.92387956
+      objectReference: {fileID: 0}
+    - target: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0.38268343
+      objectReference: {fileID: 0}
+    - target: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 45
+      objectReference: {fileID: 0}
+    - target: {fileID: 2648506453657933904, guid: 1e02f588182f44c4a97c792c640dbc44,
+        type: 3}
+      propertyPath: m_Name
+      value: PlayerMinimapIcon
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 1e02f588182f44c4a97c792c640dbc44, type: 3}
+--- !u!4 &4059791866630638263 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 1152356852073621092, guid: 1e02f588182f44c4a97c792c640dbc44,
+    type: 3}
+  m_PrefabInstance: {fileID: 4011222115521916115}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1001 &6893777000688009991
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -252,7 +328,7 @@ PrefabInstance:
     - target: {fileID: 6403528911880314253, guid: 3cc92b89ff2514a48bf0018e8d814d71,
         type: 3}
       propertyPath: m_RootOrder
-      value: 1
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 6403528911880314253, guid: 3cc92b89ff2514a48bf0018e8d814d71,
         type: 3}
