+++ Shadow/Assets/Prefabs/Monsters/AggressiveGreenSlime.prefab
@@ -289,7 +289,7 @@ Transform:
   m_Children:
   - {fileID: 7360009961570503914}
   m_Father: {fileID: 7360009962843372801}
-  m_RootOrder: 2
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!212 &6054528989439276138
 SpriteRenderer:
@@ -478,7 +478,7 @@ RectTransform:
   - {fileID: 6373211707940331188}
   - {fileID: 4203044137978957403}
   m_Father: {fileID: 7360009962843372801}
-  m_RootOrder: 3
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -822,6 +822,7 @@ Transform:
   m_Children:
   - {fileID: 3011253223712697823}
   - {fileID: 6054528990673537023}
+  - {fileID: 1942797430234210456}
   - {fileID: 6054528989439276141}
   - {fileID: 2721708261697126846}
   m_Father: {fileID: 0}
@@ -1049,3 +1050,78 @@ MonoBehaviour:
   m_hasFontAssetChanged: 0
   m_baseMaterial: {fileID: 0}
   m_maskOffset: {x: 0, y: 0, z: 0, w: 0}
+--- !u!1001 &3722409337853643407
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 7360009962843372801}
+    m_Modifications:
+    - target: {fileID: 171709150314779683, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_Name
+      value: EnemyMinimapIcon
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 2
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: fae8d86094f427d44bf4e20829600041, type: 3}
+--- !u!4 &1942797430234210456 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+    type: 3}
+  m_PrefabInstance: {fileID: 3722409337853643407}
+  m_PrefabAsset: {fileID: 0}
