+++ Shadow/Assets/Prefabs/Monsters/GoblinSlinger.prefab
@@ -365,7 +365,7 @@ RectTransform:
   - {fileID: 6025649475738953370}
   - {fileID: 3583014078929614453}
   m_Father: {fileID: 7910951102949475631}
-  m_RootOrder: 3
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -544,7 +544,7 @@ Transform:
   m_Children:
   - {fileID: 7910951103689897156}
   m_Father: {fileID: 7910951102949475631}
-  m_RootOrder: 2
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!212 &6911714829036260420
 SpriteRenderer:
@@ -665,6 +665,7 @@ Transform:
   m_Children:
   - {fileID: 2461011413946177009}
   - {fileID: 6911714828125093841}
+  - {fileID: 4309680753412629313}
   - {fileID: 6911714829036260419}
   - {fileID: 3325953626733590928}
   m_Father: {fileID: 0}
@@ -1053,3 +1054,78 @@ MonoBehaviour:
   m_hasFontAssetChanged: 0
   m_baseMaterial: {fileID: 0}
   m_maskOffset: {x: 0, y: 0, z: 0, w: 0}
+--- !u!1001 &1338021701480646998
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 7910951102949475631}
+    m_Modifications:
+    - target: {fileID: 171709150314779683, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_Name
+      value: EnemyMinimapIcon
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 2
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: fae8d86094f427d44bf4e20829600041, type: 3}
+--- !u!4 &4309680753412629313 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+    type: 3}
+  m_PrefabInstance: {fileID: 1338021701480646998}
+  m_PrefabAsset: {fileID: 0}
