+++ Shadow/Assets/Prefabs/Monsters/Minotaur.prefab
@@ -602,7 +602,7 @@ Transform:
   - {fileID: 5787899575067677586}
   - {fileID: 5787899574328366660}
   m_Father: {fileID: 7974211377366655915}
-  m_RootOrder: 2
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!212 &6667604428852898496
 SpriteRenderer:
@@ -734,7 +734,7 @@ RectTransform:
   - {fileID: 5835581717870067742}
   - {fileID: 3664279388792951025}
   m_Father: {fileID: 7974211377366655915}
-  m_RootOrder: 3
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -821,7 +821,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!4 &7974211377366655915
 Transform:
   m_ObjectHideFlags: 0
@@ -835,6 +835,7 @@ Transform:
   m_Children:
   - {fileID: 2397052125045142389}
   - {fileID: 6667604427671247189}
+  - {fileID: 4175695453567788703}
   - {fileID: 6667604428852898503}
   - {fileID: 3263816502406029076}
   m_Father: {fileID: 0}
@@ -1060,3 +1061,88 @@ MonoBehaviour:
   m_hasFontAssetChanged: 0
   m_baseMaterial: {fileID: 0}
   m_maskOffset: {x: 0, y: 0, z: 0, w: 0}
+--- !u!1001 &1201784411513998472
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 7974211377366655915}
+    m_Modifications:
+    - target: {fileID: 171709150314779683, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_Name
+      value: EnemyMinimapIcon
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 2
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalScale.x
+      value: 3
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalScale.y
+      value: 3
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: fae8d86094f427d44bf4e20829600041, type: 3}
+--- !u!4 &4175695453567788703 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+    type: 3}
+  m_PrefabInstance: {fileID: 1201784411513998472}
+  m_PrefabAsset: {fileID: 0}
