+++ Shadow/Assets/Prefabs/Monsters/Skeleton.prefab
@@ -196,6 +196,7 @@ Transform:
   m_Children:
   - {fileID: 5578080629364511097}
   - {fileID: 3505293736718498649}
+  - {fileID: 4432054113153687909}
   - {fileID: 3505293735471761611}
   - {fileID: 4712523481413287192}
   m_Father: {fileID: 0}
@@ -455,7 +456,7 @@ Transform:
   m_Children:
   - {fileID: 182200130480926796}
   m_Father: {fileID: 182200131758195111}
-  m_RootOrder: 2
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!212 &3505293735471761612
 SpriteRenderer:
@@ -654,7 +655,7 @@ RectTransform:
   - {fileID: 4382422816411329042}
   - {fileID: 6843072586765660925}
   m_Father: {fileID: 182200131758195111}
-  m_RootOrder: 3
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -1054,3 +1055,78 @@ MonoBehaviour:
   m_OnValueChanged:
     m_PersistentCalls:
       m_Calls: []
+--- !u!1001 &1503997568394087282
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 182200131758195111}
+    m_Modifications:
+    - target: {fileID: 171709150314779683, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_Name
+      value: EnemyMinimapIcon
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 2
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: fae8d86094f427d44bf4e20829600041, type: 3}
+--- !u!4 &4432054113153687909 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 2980983187309324823, guid: fae8d86094f427d44bf4e20829600041,
+    type: 3}
+  m_PrefabInstance: {fileID: 1503997568394087282}
+  m_PrefabAsset: {fileID: 0}
