+++ Shadow/Assets/Prefabs/UI/Minimap.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b80714a94f2d4c240b6bcbb6dfc922b8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
