+++ Shadow/Assets/Prefabs/UI/Minimap/EnemyMinimapIcon.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fae8d86094f427d44bf4e20829600041
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
