+++ Shadow/Assets/Prefabs/UI/Minimap/MinimapCamera.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e4f3e22f8d18d5b45b6cd5766a6bc732
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
