+++ Shadow/Assets/Prefabs/UI/Minimap/MinimapCanvas.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5778719eed04a454c90bec20f4085fea
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
