+++ Shadow/Assets/Prefabs/UI/Minimap/Minimap_QuestAvailable.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0a6b1318f15e4644ea1f6871cb0deed7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
