+++ Shadow/Assets/Prefabs/UI/Minimap/Minimap_QuestComplete.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 84faf31d0b006384089c78aaa299edc0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
