+++ Shadow/Assets/Prefabs/UI/Minimap/NPCMinimapIcon.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 310e8405a2f75ec4bb90204c1aa3cb29
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
