+++ Shadow/Assets/Prefabs/UI/Minimap/PlayerMinimapIcon.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1e02f588182f44c4a97c792c640dbc44
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
