--- Shadow/Assets/Prefabs/UI/Quests/Minimap_QuestAvailable.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0a6b1318f15e4644ea1f6871cb0deed7
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
