--- Shadow/Assets/Prefabs/UI/Quests/Minimap_QuestComplete.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 84faf31d0b006384089c78aaa299edc0
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
