+++ Shadow/Assets/Scenes/Loading.unity
@@ -123,6 +123,125 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
+--- !u!1001 &129244898
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_Pivot.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_Pivot.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 12
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_AnchorMax.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_AnchorMax.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_AnchorMin.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_AnchorMin.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_SizeDelta.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_SizeDelta.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919473, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1487316800370919477, guid: 5778719eed04a454c90bec20f4085fea,
+        type: 3}
+      propertyPath: m_Name
+      value: MinimapCanvas
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 5778719eed04a454c90bec20f4085fea, type: 3}
 --- !u!1001 &371382885
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -391,6 +510,75 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: fac8c40b9e6eeda4799617b378e885ea, type: 3}
+--- !u!1001 &483977417
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 2369654402055757987, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 11
+      objectReference: {fileID: 0}
+    - target: {fileID: 2369654402055757987, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -5.8
+      objectReference: {fileID: 0}
+    - target: {fileID: 2369654402055757987, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -23.04
+      objectReference: {fileID: 0}
+    - target: {fileID: 2369654402055757987, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2369654402055757987, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2369654402055757987, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2369654402055757987, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2369654402055757987, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2369654402055757987, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2369654402055757987, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2369654402055757987, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2369654402055757998, guid: e4f3e22f8d18d5b45b6cd5766a6bc732,
+        type: 3}
+      propertyPath: m_Name
+      value: MinimapCamera
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: e4f3e22f8d18d5b45b6cd5766a6bc732, type: 3}
 --- !u!1 &733388585
 GameObject:
   m_ObjectHideFlags: 0
