--- Shadow/Assets/Scripts/Minimap.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b16f78e5dda6a9d4da309da462beed88
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
