+++ Shadow/Assets/Scripts/Singletons/MinimapWindow.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MinimapWindow : Singleton<MinimapWindow>
+{
+
+}
