+++ Shadow/Assets/Scripts/Singletons/MinimapWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0d81c6da3c2ba6240a781c15526ef058
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
