+++ Shadow/Assets/Scripts/StoryEvents/AppearWhileScenario.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 42ce35ad2152cba4fa3f0243af18ebfa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
